+++ src/print.c
@@ -31,7 +31,7 @@ void print_type(const struct Type* type) {
         case ContType: {
             printf("cont (");
             const struct Types *params = &type->payload.cont.param_types;
-            for (int i = 0; i < params->count; i++) {
+            for (size_t i = 0; i < params->count; i++) {
                 print_type(params->types[i]);
                 if (i < params->count - 1)
                     printf(", ");
@@ -41,7 +41,7 @@ void print_type(const struct Type* type) {
         } case FnType: {
             printf("fn (");
             const struct Types *params = &type->payload.fn.param_types;
-            for (int i = 0; i < params->count; i++) {
+            for (size_t i = 0; i < params->count; i++) {
                 print_type(params->types[i]);
                 if (i < params->count - 1)
                     printf(", ");
